refactor: move resources to testclasses in staticFieldAccess2
