--- src/test/java/spoon/test/staticFieldAccess2/ALong.java
@@ -1,6 +1,0 @@
-package spoon.test.staticFieldAccess2;
-
-public @interface ALong
-{
-	public long number() default 7;
-}
