--- src/test/java/spoon/test/staticFieldAccess2/Constants.java
@@ -1,6 +1,0 @@
-package spoon.test.staticFieldAccess2;
-
-public class Constants
-{
-	public static final long PRIO = 1;
-}
