--- src/test/java/spoon/test/staticFieldAccess2/ImplicitFieldReference.java
@@ -1,20 +1,0 @@
-package spoon.test.staticFieldAccess2;
-
-public class ImplicitFieldReference {
-	public String memberField;
-
-	public String getMemberField() {
-		return memberField;
-	}
-
-	public void setMemberField(String p_memberField) {
-		memberField = p_memberField;
-	}
-	public void setMemberField2(String memberField) {
-		this.memberField = memberField;
-	}
-	
-	public void testLocalMethodInvocations() {
-		getMemberField();
-	}
-}
