+++ src/test/java/spoon/test/staticFieldAccess2/testclasses/ALong.java
@@ -1,0 +1,6 @@
+package spoon.test.staticFieldAccess2.testclasses;
+
+public @interface ALong
+{
+	public long number() default 7;
+}
