+++ src/test/java/spoon/test/staticFieldAccess2/testclasses/Constants.java
@@ -1,0 +1,6 @@
+package spoon.test.staticFieldAccess2.testclasses;
+
+public class Constants
+{
+	public static final long PRIO = 1;
+}
