+++ src/test/java/spoon/test/staticFieldAccess2/testclasses/ImplicitStaticClassAccess.java
@@ -1,0 +1,18 @@
+package spoon.test.staticFieldAccess2.testclasses;
+
+public class ImplicitStaticClassAccess {
+	static String ImplicitStaticClassAccess = "";
+	
+	ImplicitStaticClassAccess() {
+		this(ImplicitStaticClassAccess.class);
+	}
+	
+	ImplicitStaticClassAccess(Class<?> clazz) {
+		ImplicitStaticClassAccess.class.getName();
+	}
+
+	public void testLocalMethodInvocations() {
+		ImplicitStaticClassAccess.class.getName();
+	}
+	
+}
